relace existing bookmarks test with revised test
